backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-cloud-aiplatform==1.49.0
vertexai==1.49.0
requests==2.32.3
python-dotenv==1.0.1
mem0ai==0.1.29
chromadb==0.5.3
pydantic==2.9.2
